requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
Flask==2.0.1
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.3.1
pyTelegramBotAPI==3.8.2
requests==2.26.0
urllib3==1.26.6
Werkzeug==2.0.1
